requirements.txt
requests
urllib3
dnspython
psutil
